+++ partiql-common/src/node.rs
@@ -1,5 +1,4 @@
 use indexmap::IndexMap;
-use std::hash::Hash;
 use std::sync::{Arc, RwLock};
 
 #[cfg(feature = "serde")]